"rust-analyzer.debug" command
