+++ editors/code/package.json
@@ -121,6 +121,11 @@
                 "category": "Rust Analyzer"
             },
             {
+                "command": "rust-analyzer.debug",
+                "title": "Debug",
+                "category": "Rust Analyzer"
+            },
+            {
                 "command": "rust-analyzer.analyzerStatus",
                 "title": "Status",
                 "category": "Rust Analyzer"
